TS-1470: Fix cache sizes > 16TB (part 2)
The first part allowed you two create a cache larger than 16TB,
this second part fixes a subsequent bug that clears a large
cache on restart.

backport of: 61496440b65ee28c93bf77fa036c0424b33dc10c
test/review: igalic, arno, zwoop
backport: sorber

+++ iocore/cache/Cache.cc
@@ -2241,7 +2241,7 @@ cplist_update()
     ConfigVol *config_vol = config_volumes.cp_queue.head;
     for (; config_vol; config_vol = config_vol->link.next) {
       if (config_vol->number == cp->vol_number) {
-        int size_in_blocks = config_vol->size << (20 - STORE_BLOCK_SHIFT);
+        off_t size_in_blocks = config_vol->size << (20 - STORE_BLOCK_SHIFT);
         if ((cp->size <= size_in_blocks) && (cp->scheme == config_vol->scheme)) {
           config_vol->cachep = cp;
         } else {
